requirements.txt
dbt-core==1.9.2
dbt-duckdb==1.9.1
deltalake==0.24.0